Replicated Test Examples for Objects and Thread
I copied the code from the Digital Ocean website to test out how
the wait() and notify() method call of objects affect the performance
of threads. I Found out that wait is similar to the sleep method, both
methods invoke a process of stopping the operation of the thread. And the
difference lies in the condition for re-activating the thread. The wait method
waits for the notify to start the thread that currently is stopped by the object
through a syncronized block, I'm not sure of yet how it works. Though the sleep method
works by waiting for the time input to end.